Upgrade rclone and base image
This commit bumps the packaged rclone version to v1.50.2. The changelog
for this release is available (here)[1]. The major changes (to
rclone-1.49.1) are addressing several memory leaks and more robustness.

This commit also bumps the base image version to Alpine-3.11.3.

In addition, the Dockerfile now uses opencontainers.org labels to
annotate the created container image.

[1]: https://github.com/rclone/rclone/blob/master/docs/content/changelog.md#v1502---2019-11-19